Fix zigzag ironing pattern not reaching the walls with 0 inset
When the zigzag ironing pattern was selected, even with ironing inset at 0, the
ironing pattern was not reaching the walls. This happened because the infill
generator, which is used for generating the ironing, assumes that the pattern should
reach up to the outer wall. This commit fixes that by compensating for the outline_offset
before generating the ironing through the infill generator.

CURA-7318