Fix concentric ironing pattern exceeding the walls at 0 ironing inset
When the concentric ironing pattern was selected, the ironing pattern exceeds the walls,
leading to unnecessary travels. This happened because the infill generator, which is used
for generating the ironing, adds an extra line_width / 2 to the outline_offset.
This commit fixes that by compensating for the outline_offset before generating the ironing
using the infill generator.

CURA-7318