Also adjust support infill layer thickness through flow
Just like 545165f010a57bbcb61323d5b0f4b579ca73706e and for the same reason.
Only here there is no effect since the line width of this config is not used for the compensating offset.

I suppose that's a separate bug. It's not keeping the same centre line but aligning it to the outside of the support area.

Contributes to issue CURA-7494.