Move index-based ordering to InsetOrderOptimizer too
And it compiles again. Important.
This shouldn't change anything, but it did. It fixes the bug that insets are printed with an additional outer wall on the inside too. It also introduces a bug somehow that the last segment of a path is no longer printed. Strange.

Contributes to issue CURA-7554.