Ignore the extra skin wall count in concentric
When we set Bottom Layer Initial Pattern or top/bottom Pattern to concentric, the extra skin wall
count setting doesn't have any effect. In such cases, no insets should be generated in the skin,
except if that skin is a Top Surface (roofing) skin.

This commit fixes that by not generating insets for the layers where the pattern is concentric.
If the layer is a roofing layer, then if no insets have been generated previously, generate them
now, even if the Top Surface Skin Pattern is concentric, and take into account the Extra Skin Wall
Count to generate extra walls (if needed).

CURA-6684