Fix producing leftover unsupported roof parts
My theory:

TreeSupport generates the support and roof layer for each layer_nr and then it subtracts the X/Y
offsetted model from those two to make sure there is no collision with it. This works fine for the
support_layer, since it is a closed circle, so the difference with 0 radius eliminates the circle
completely.
For the roof_layer on the other hand, which will not be a closed circle when printed, but rather
only a pattern without an outline, when we take the difference of it from the model without using
the branch radius, the difference leaves the parts of the infill line that are not colliding with
the x/y offsetted model. These are the exact leftover parts that seem unsupported. In order to
correctly remove these too, we need to take the difference from the X/Y offsetted model, taking
into account the branch radius (as it is the maximum distance from the x/y offsetted model that
the support roof infill lines can reach).

CURA-7168