Group indexed-ordering technique by inset to print per group
This way the order optimizer can optimise them together to reduce travel moves.
However that order optimizer doesn't seem to work properly right now. It doesn't deal with polylines but with polygons. It also tends to go back and forth along the same line a lot, and produces lines going into infinity. That needs to be fixed.

Contributes to issue CURA-7558.