Remove GcodeWriter class
This class was duplicating the functionality of FffGcodeWriter. We had previously left this in as a reference of how to extrude lines. It looks like we'll need it less, now. The approach that it was using to optimize lines was useful, but no longer valid since it needs to be refactored to deal with the seam.
As I'm refactoring this, it was breaking compilation. Rather than fixing the dead code, I'm removing it here. If you need it as a reference, see the original pull request from BagelOrb.

Contributes to issue CURA-7558.