Use a template for the input data of the order optimizer
This is necessary because some consumers use the order optimizer to determine the order of parts in a completely different list than what was actually given to the optimizer. They did that because the optimizer could only handle polygons, but they actually needed to have various different things ordered (SkinPart, SupportInfillPart, etc.) Since I removed that indirection they now need to provide the actual data they want to optimize. No more working via indices. However that needs a template.
Because it's a template, all of the implementation goes back into the header, even though I just moved it. D'oh!

Contributes to issue CURA-7558.