Directly use type for the order optimiser paths rather than pointers
This is necessary because the Polygons class creates ConstPolygonRef instances upon using the [] operator. So when we're reordering these, we get a new instance every time we're adding a ConstPolygonRef, which then immediately gets deallocated while we only store the pointer in the optimiser. So we're just storing addresses that have been invalidated by the time we're able to optimise, and by the time the consumer needs the optimised paths back.
Instead, the consumer now needs to provide pointers to optimise if they want to prevent copies. This way they can choose to still optimise ConstPolygonRefs (but those will simply make copies of the reference, which is not expensive). If they're using a more expensive data structure, such as SkinPart, they will need to provide pointers to them and specialise the getVertexData function to accept pointers.

Contributes to issue CURA-7558.