Implement loop detection on polylines
The polylines will then be turned into polygons if they are looping. This is necessary because libArachne sometimes produces polylines which are actually loops. In those cases we need to close them and treat them as polygons for the remainder of the optimization. That way we'll be able to place the seam correctly even if the lines are added as polylines.

Contributes to issue CURA-7558.