Still find the closest vertex if none of them match the filter in user-specified seams
If the seam type is user-specified, and you've set the corner preference to allow only inner corners or outer corners, but none of the vertices match that preference, still choose the closest vertex to the desired seam position. Previously all those scores would end up on std::numeric_limits<float>::max(), which made them all equal and it would choose an arbitrary vertex. Now it still has those ordered.
I'm now using the actual vSize rather than vSize2. Otherwise this 1 metre penalty would bring the score up in the ranges of 10^12, which makes the floating point rounding errors on the order of 4096. This should be much more precise.

Contributes to issue CURA-7558.