Use vectors instead of unordered_sets for the contact_nodes
In every slice, a different MinimumSpanningTree was generated every time due to the fact that
Prim's algorithm was receiving an unordered set of vertices, thus accessing the vertices in
different order, every time it was run. Now, since the contact points are generated in a grid,
they generally have the same distane between them. So, depending on which neighbour was found
first, the MST would generate different edges.
As the nodes move down, this difference becomes more and more apparent, because of the different
node merges, thus leading in an almost entirely different tree in every slice.

By converting the unordered sets into vectors, the vertices are accessed in the same order, so
the non-deterministic behaviour is eliminated.

The performance should be impacted slightly when there are unsupported branch leaves (since then
the TreeSupport is searching to find a specific node in a vector) or when the insertDropNode is
searching if a node already exists in the layer.

CURA-7547