Fix references to vertex cache being invalidated
When adding a new polygon or polyline, references to the cache would get invalidated because the cache vector gets resized. So to fix that, we can only store them in the cache once we know that the cache is not getting resized any more; when the optimize() function is called.

Contributes to issue CURA-7558.