Optionally use combing distance to check shortest paths
Only compute it if absolutely necessary though.
Because of the check against the direct distance, I expect that this will be much faster than the original code when checking against the combing distances. It now earlies out if the direct distance is already longer than the combing distance of the best candidate.

Contributes to issue CURA-7558.