Fix combining infill layers
In commit 17f5365ba49857047fbb899ca83409c2ef09e710 the Spaghetti Infill routine was removed here, for which the indentation was changed. There was a mistake there with how the brackets were read. This line should also have been unindented. It belongs to the normal infill generation routine, not to the Sierpinski fractal pre-computation.

Fixes CURA-7613.