Fix outer wall on inside.
Using a meta-strategy (actually a convenience class that seems to combine both the meta and underlying strat) to ignore walls beyond the count. This enables us to use the whole shape instaed of first cutting out a tube/donut and _then_ applying lib-arachne. Besides the 'outer wall also in inside' this will fix a few other problems, like the small gaps we sometimes saw in the 'middle' wall and maybe the problems we have with getting support walls to be a proper ammount.

CURA-7553