Fix test of reading mesh group
We need to provide it with infill_mesh, cutting_mesh and anti_overhang_mesh values since that's necessary for min() and max() of the mesh group. We also need to adjust the global default for material_shrinkage_percentage since a value of 0 now makes it resize the mesh to 0.

Contributes to issue CURA-7118.