Remove code that moved tree support into the center of the model
I couldn't find any reason for it to be there. The documentation didn't mention
why it was happening. Only something wrt to tree support needing to be supported
on diagonal walls lived in some collective memory bits. Removing this seems to
leave that behavior on all the while fixing the issue of tree support being
removed unneeded

CURA-7668