Added a class for WallToolPath generation
To avoid a lot of code duplication a class WallToolPaths was created.
This class is responsible for the creation of libArachne type walls.
It will also hold various helper functions, such as the conversion from
toolpaths to a bin of junctions or obtaining the inner contour from
toolpaths for now it returns a simple offset but in the future it
seams like a good place to implement CURA-7681.

CURA-7626_infill_walls_with_libArachne