Fix InwardDistributed (without messing up even Distributed).
The previous number of commits in this branch where made on the assumption that the way the Cura-frontend displays line segment-width is always correct. This caused me to see bug where there where none and attempt to fix those. The result is that I went over everything _again_ with a fine toothed comb, but now without that assumption. This is hopefully the last fix in this ticket. My apologies for the confusing git history ;-)   ... I _think_ the calculations are at least correct according to the paper now, but I'm not entirely sure this will always result in ok prints, as the line-width may be too wide to print in some cases now, as it should always be prepared to print 1+1/2 line width in case of xxDistributed as Distributed doesnt have any thickness left over and will always result in an even number of walls if the part is thin enough not to need skin.

part of CURA-7676