Finally fix xDistributed.
The thing that continued to be wrong in the previous commits is that, if the outer wall width is taken as a given, then there is one less wall to distribute everything else over. When you test with something that only takes 2 walls max (on one side) the reduced flexibility of the situation sometimes leads to missing walls in the middel (since the distributed strategies usually don't have a wall in the middle, unless there's only 'half' a wall on each side so to speak. Now, instead of keeping the outer wall always at the same width (unless it was the only one), now only the _realtive_ width between the differing widths (outer and inner) is preserved in the distributed schemes.

part of CURA-7676