Changed asserts  for crossing on a line to catch unwanted behaviour
A crash can occur when applying an odd infill multiplier See comments
 in ticket CURA-7626 29/09/2020.

A sigsev occurs due to the vertex_index being out of range. Not quite
sure why. Previous crashes in this code region was due to the fact that
the whole crossings_on_line was empty. Which occurred if the outline
was too small for some reason.

CURA-7626_infill_walls