Remove intermediate vertexes in really long line segments
Due to rounding errors when calculating the sqrt(ab_size2), simplify was still leaving vertexes
that existed in very long line segments. To get around that issue, getDistFromLine is introduced
that calculates that distance following a different formula than the dot product. This way, it
avoids calculating the squares and square roots of enormous lines that would otherwise generate
overflow and rounding errors.

This getDistFromLine is now used inside simplify when deciding to remove this vertex, thus making
simplify less sensitive to overflows.

CURA-7769