Don't do full optimization, instead offer choice to print by insets or regions.
Attempted to build optimization on top of the insets from libArachne. Didn't work out and it turns out libArachne output is close to desired behaviour anyway. Decided to keep the part that emulated the old optimize behaviour, but switched around the defaults, since the 'emulate old optimize' behaviour will often be slower (and worse if there are a lot of small segments between regions, since its split up by region -> this is a known issue now and one of the smaller reasons it was decided not to pursue optimization in this case).

CURA-7557 CURA-7677