Remove outline_offset parameter from Infill constructor
It was 0 everywhere except in the ironing generation. In the ironing code I've just applied it to the area that needs to be ironed instead.
This parameter needs to be removed since it's involved in the offset for the infill pattern within the extra infill walls. That offset is no longer valid if you're using variable-width infill lines.

Contributes to issue CURA-7681.