Cut previous generated denser infills with the new infill as tool
Hand over commit, not fully functional yet. The working idea was that
the previous generated polygons and lines are the be cut with the earlier
generated polygons and or lines. Leaving the spare infill area untouched
and connected. While subsequently generated denser infill areas are cut
where there are already existing lines.

This approach seems to work for Polygons generated with the infill line
multiplier of values of 3 and lower. Once these go above the tool Polygon
in the lambda function cut_polygon is incorrect, I believe it needs to
only use the most inner and outer polygons in all cases. But I didn't
came round to that.

I also adjusted the way how the lines are cut. In earlier commits the
denser infill areas were cut with an offset from the spare infill contour
This would still result in over extrusion in the crossings of the infill
lines. The lambda cut_lines also doesn't function yet.

As a final remark I want to say: "stupid polygons!"

Contributes to CURA-7115