Workaround: add a 0-width wall in locations where there are maximum walls
This is a workaround for an issue where it's missing a junction for an odd-numbered inner wall. This will make a double wall with a few micron spacing (rounding errors?) in places where there is exactly the maximum number of walls. These very thin infill areas are easily filtered.
Once the odd-numbered inner walls are generated properly again, we may be able to remove this workaround.

Contributes to issue CURA-7681.