Stitch walls to closed loops as much as possible
Not all 0-length walls are closed loops due to the way in which they are generated. If two opposite walls come very close together such that the inner walls merge together, the innermost walls will not be attached and form a series of polylines that together close a loop. We must connect those polylines together to form a loop.
It must stitch by up to half a line width due to a bug where singular odd walls are not always generated.

Contributes to issue CURA-7681.