Don't pre-filter on the amount of insets
The amount of insets is unreliable since the insets are not going to determine the actual shape of the walls with variable line width.
If the inner area is truly nothing, it'll end up with an empty polygon and all operations on that will be very quick anyway.

Contributes to issue CURA-7681.