Made a meta-beading-strat. that takes outer and inner wall widths into account.
The idea is as follows: Initially, for the parent (base) beading strategy, set the 'optimum' width as a weighted average defined by 'what the total thickness of the fill would be if max-bead width is always reached' (so take the situation where the actually requested total wall width is reached as the ground truth we base all the rest on). The base beading-strategy will thus produce the 'right' number of walls but the wrong bead-widths (and locations) -- becasue it assumes the walls are some average. This is then rectified by the newly introduced meta-strategy, which redistributes the bead-widths. It takes a look at the factor of the thickness the outer and inner walls would have occupied in the ideal situation and applies that to the current situation. This keeps a) the original beadings simple to understand and debug b) the meanings and behaviour of the original beadings intact while still c) preserving the widths of the outer and inner walls the user (settings) has/have input at the very least in the case of 'full' walls (and does a decent attempt at at least preserving the relative relation between those two values with a lesser thickness).

part of CURA-7597