Don't remove junctions with distinctively different widths
In some cases, the wall generation introduces junctions which are used to change the width of the
wall, in the middle of a straight line. Generally, these junctions won't have huge difference in
the new width so it is safe to remove them. However, if Print Thin Walls is enabled and the
widening strategy is activated, there may be a big difference of width. When that happens, simplify
will not remove these junctions, if the width change is more than 20% of the nominal wall width.

CURA-7751