Remove junctions of straight lines based on a setting in simplify
A new meshfix setting named "Maximum Area Deviation Error" is introduced. This setting denotes
what is the maximum allowed area lost or gained when the middle junction of a straight line
is removed. This area deviation exists due to the difference of width of the colinear segments.

A             B                          C              A                                        C
---------------                                         ************** deviation_error / 2
|             |---------------------------  B removed   ------------------------------------------
|             |                          |  --------->  |            |                           |
|             |---------------------------              ------------------------------------------
---------------             ^                           **************     ^
      ^                    C.w                                            C.w
     B.w

CURA-7751