Actually use infill pattern generator to generate extra walls
Previously we were using the infill pattern generator with a hard-coded number of walls of 0. The walls for the skin were generated separately. This required keeping track of an additional polygon for where the infill pattern of the skin would go, and a specific call to libArachne to generate these walls. Let's not do that. It was also causing problems because this extra call was not taking the skin overlap into account.
Let's just use the infill pattern generator with an appropriate number of extra infill walls. Use the skin configs to generate them so that the extra skin wall gets the appropriate width. The infill pattern generator generates its pattern within the area that is properly expanded with the skin overlap. And presto.

Only the top surface skin feature has broken. I'm not sure if I just broke it, but I touched it anyway. I need to look at that feature again now. This skin/infill generation code is a big mess of spaghetti.

Contributes to issue CURA-7681.